simple line drawing

+++ include/Canvas.hpp
@@ -15,7 +15,7 @@ public:
     void fill(const Col4& col);
     
     // draw a line
-    void line(const Vec2f& start, const Vec2f& end, const Col4& col);
+    void line(const Vec2i& start, const Vec2i& end, const Col4& col);
 };
 
 NS_SPECTRUM_END
